No need to remove licence header anymore
Works on the assumption that the header always starts with #.
There is a loop that checks if the line starts with #. If it does, it passes to the next line. If it doesn't, it means that it reached the empty line between the license header and  the dictionary. The empty line is ignored and the loop is finished.